🐛 fix:permissions.py 분리, patch메서드 post메서드로 변경에 따른 urls변경, config에 studies 인클루드"
# Conflicts:
#	config/urls.py
