+++ apps/studies/serializers/members.py
@@ -1,7 +1,7 @@
 from rest_framework import serializers
 
 
-class LeaderDelegationSerializer(serializers.Serializer):  # type: ignore[type-arg]
+class DelegateLeaderSerializer(serializers.Serializer):  # type: ignore[type-arg]
     """REQ-STDY-006: 리더 위임 요청 시 사용되는 입력값 검증"""
 
     target_user_id = serializers.IntegerField(required=True, help_text="리더로 위임할 대상 멤버의 유저 ID")
